Refactor http_client_impl.h to be internal.
