+++ Release/include/cpprest/http_headers.h
@@ -328,4 +328,18 @@ private:
     std::map<utility::string_t, utility::string_t, _case_insensitive_cmp> m_headers;
 };
 
+namespace details {
+
+    /// <summary>
+    /// Serialize the http_headers into name:value pairs separated by a carriage return and line feed.
+    /// </summary>
+    utility::string_t flatten_http_headers(const http_headers &headers);
+#if defined(_WIN32)
+    /// <summary>
+    /// Parses a string containing Http headers.
+    /// </summary>
+    void parse_headers_string(_Inout_z_ utf16char *headersStr, http_headers &headers);
+#endif
+}
+
 }}
