+++ Release/src/CMakeLists.txt
@@ -4,6 +4,7 @@ if (NOT CPPREST_EXCLUDE_WEBSOCKETS)
 endif()
 
 set(SOURCES_COMMON
+  http/client/http_client.cpp
   http/client/http_client_msg.cpp
   http/client/x509_cert_utilities.cpp
   http/common/http_helpers.cpp
