+++ Release/src/http/client/http_client_asio.cpp
@@ -43,7 +43,7 @@
 #error "Cpp rest SDK requires c++11 smart pointer support from boost"
 #endif
 
-#include "cpprest/details/http_client_impl.h"
+#include "http_client_impl.h"
 #include "cpprest/details/x509_cert_utilities.h"
 #include <unordered_set>
